refactor: move things around
