+++ assets/scripts/segments/constants.js
@@ -10,10 +10,18 @@ export const MAX_SEGMENT_WIDTH = 400
 
 export const DRAGGING_MOVE_HOLE_WIDTH = 40 // pixels
 
+export const DRAGGING_TYPE_NONE = 0
+export const DRAGGING_TYPE_MOVE = 2
+export const DRAGGING_TYPE_RESIZE = 3
+
 export const SEGMENT_WARNING_OUTSIDE = 1
 export const SEGMENT_WARNING_WIDTH_TOO_SMALL = 2
 export const SEGMENT_WARNING_WIDTH_TOO_LARGE = 3
 
-export const DRAGGING_TYPE_NONE = 0
-export const DRAGGING_TYPE_MOVE = 2
-export const DRAGGING_TYPE_RESIZE = 3
+export const SEGMENT_WIDTH_RESOLUTION_IMPERIAL = 0.25
+export const SEGMENT_WIDTH_CLICK_INCREMENT_IMPERIAL = 0.5
+export const SEGMENT_WIDTH_DRAGGING_RESOLUTION_IMPERIAL = 0.5
+
+export const SEGMENT_WIDTH_RESOLUTION_METRIC = 1 / 6
+export const SEGMENT_WIDTH_CLICK_INCREMENT_METRIC = 2 / 6
+export const SEGMENT_WIDTH_DRAGGING_RESOLUTION_METRIC = 2 / 6
