+++ assets/scripts/users/constants.js
@@ -1,10 +1,2 @@
 export const SETTINGS_UNITS_IMPERIAL = 1
 export const SETTINGS_UNITS_METRIC = 2
-
-export const SEGMENT_WIDTH_RESOLUTION_IMPERIAL = 0.25
-export const SEGMENT_WIDTH_CLICK_INCREMENT_IMPERIAL = 0.5
-export const SEGMENT_WIDTH_DRAGGING_RESOLUTION_IMPERIAL = 0.5
-
-export const SEGMENT_WIDTH_RESOLUTION_METRIC = 1 / 6
-export const SEGMENT_WIDTH_CLICK_INCREMENT_METRIC = 2 / 6
-export const SEGMENT_WIDTH_DRAGGING_RESOLUTION_METRIC = 2 / 6
